requirements.txt
requests~=2.27.1
oauthlib~=3.2.0
requests-oauthlib~=1.3.1
python-telegram-bot~=13.11
